Add setting in rk_abi.module to hide Kõik link on homepage.
The views slideshow at the top of the homepage has pager
controls, but one always displays and the ministries just
do not want to see that link there. Now there is a new
setting in the rk_abi module to hide that link. When the
settings is checked, then additional css rules are loaded
to remove the background arrow and do visibility:hidden
on the actual text. (keep it there for accessibility instead
of display:none;)